refactor: simplify routing
