+++ frontend/src/App.tsx
@@ -1,15 +1,8 @@
 import { useEffect } from "react";
 import { Routes, Route } from "react-router-dom";
 import { getAggregatedDates, getSites, getStopPoints } from "./store/actions";
-import { useAppDispatch, useAppSelector } from "./store/store";
+import { useAppDispatch } from "./store/store";
 import { SidebarPresenter } from "./presenters/sidebarPresenter";
-import {
-    getSitesLoadingCB,
-    getStopPointsCB,
-    getStopPointsLoadingCB,
-    getSitesCB,
-} from "./store/selectors";
-import { Suspense } from "./components/Suspense";
 import { ROUTES, type RouteConfig } from "./routes";
 
 function App() {
@@ -21,16 +14,6 @@ function App() {
         dispatch(getStopPoints());
         dispatch(getAggregatedDates());
     }, [dispatch]);
-
-    const sites = useAppSelector(getSitesCB);
-    const isSitesLoading = useAppSelector(getSitesLoadingCB);
-    const stopPoints = useAppSelector(getStopPointsCB);
-    const isStopPointsLoading = useAppSelector(getStopPointsLoadingCB);
-
-    // TODO: this loading should probably only be checked when on the map page
-    if (isSitesLoading || !sites || isStopPointsLoading || !stopPoints) {
-        return <Suspense fullscreen message="Loading transit data and preparing the map..." />;
-    }
 
     function renderRouteCB(route: RouteConfig) {
         return <Route key={route.path} path={route.path} element={route.element} />;
